+++ src/Firebase/Contract/Auth.php
@@ -379,26 +379,6 @@ interface Auth
     public function signInAnonymously(): SignInResult;
 
     /**
-     * @deprecated 5.26.0 Use {@see signInWithIdpAccessToken()} with 'twitter.com' instead.
-     */
-    public function signInWithTwitterOauthCredential(string $accessToken, string $oauthTokenSecret, ?string $redirectUrl = null, ?string $linkingIdToken = null): SignInResult;
-
-    /**
-     * @deprecated 5.26.0 Use {@see signInWithIdpIdToken()} with 'google.com' instead.
-     */
-    public function signInWithGoogleIdToken(string $idToken, ?string $redirectUrl = null, ?string $linkingIdToken = null): SignInResult;
-
-    /**
-     * @deprecated 5.26.0 Use {@see signInWithIdpAccessToken()} with 'facebook.com' instead.
-     */
-    public function signInWithFacebookAccessToken(string $accessToken, ?string $redirectUrl = null, ?string $linkingIdToken = null): SignInResult;
-
-    /**
-     * @deprecated 5.26.0 Use {@see signInWithIdpIdToken()} with 'apple.com' instead.
-     */
-    public function signInWithAppleIdToken(string $idToken, ?string $rawNonce = null, ?string $redirectUrl = null, ?string $linkingIdToken = null): SignInResult;
-
-    /**
      * @see https://cloud.google.com/identity-platform/docs/reference/rest/v1/accounts/signInWithIdp
      *
      * @param \Stringable|string $provider