+++ CHANGELOG.md
@@ -41,6 +41,12 @@ methods.
 * `Kreait\Firebase\Contract\Auth::confirmPasswordReset()` now returns the email address the password reset was confirmed for.
     * This was previously done with the method `Kreait\Firebase\Contract\Auth::confirmPasswordResetAndReturnEmail()` -
       this method has been removed.
+* The following methods were shortcuts for `Kreait\Firebase\Contract\Auth::signInWithIdpAccessToken()` and
+  `Kreait\Firebase\Contract\Auth::signInWithIdpIdToken()` and have been removed.
+  * `Kreait\Firebase\Contract\Auth::signInWithAppleIdToken()`, use `signInWithIdpIdToken('apple.com', ...)` instead
+  * `Kreait\Firebase\Contract\Auth::signInWithFacebookAccessToken()`, use `signInWithIdpAccessToken('facebook.com', ...)` instead
+  * `Kreait\Firebase\Contract\Auth::signInWithGoogleIdToken()`, use `signInWithIdpIdToken('google.com', ...)` instead
+  * `Kreait\Firebase\Contract\Auth::signInWithTwitterOauthCredential()`, use `signInWithIdpAccessToken('twitter.com', ...)` instead
 * The following methods now return strings instead of value objects:
     * `Kreait\Firebase\Contract\Auth::confirmPasswordReset()`
     * `Kreait\Firebase\Contract\Auth::verifyPasswordResetCode()`